Add test support
